Bot KPlayer: module KPlayerBot + 4 cua giao tiep (moc KHONG HOI QUY dau tien)
Buoc 1-4 cua ke hoach dot "goi duoc bot". Chua bat cai gi - moi thu con ngu.

=== Don nốt bay bo cuc (Buoc 1) ===
Ban fork Sources\Core\Src\KProtocol.h con 3 truong size_t nua ma ban Headers khai la
unsigned int (grep size_t tren Headers = 0): tagDBSyncPlayerInfo:1342, tagGuidableInfo:1370,
tagGS2GWSaveRole:1611. Cung loai bay voi TProcessData - tren x64 thanh 8 byte trong khi
Goddess/Bishop (Win32) doc 4. Da doi ca ba.

=== KPlayerBot.h / .cpp (moi) ===
Bot = KPlayer THAT nap tu roledb, KHAC HAN bot NPC cua KSimCity (KNpc kind_player khong co
KPlayer). Dat tien to PB_ de khong lan voi SC_ / m_btSimCityBot.

NGUYEN TAC KIEN TRUC: CORE KHONG BAO GIO TU DUNG BYTE LEN DAY.
Cay nay co hai KProtocol.h chung guard; neu de Core dong goi TProcessData roi ban sang
Goddess thi Goddess doc nham ulIdentity, ten tai khoan lech 4 byte, GetRoleListOfAccount
that bai MA KHONG CO MOT DONG BAO LOI NAO. Nen Core chi dua (viec gi, ulIdentity, ten) qua
con tro ham PB_DbSender; GameServer moi dung goi tin bang chinh khoi ma da chay 20 nam
(KSOServer.cpp:3193-3211). Bay khong the xay ra, khong phu thuoc sizeof.

Giao thuc da doc tan noi de bo cuc dung:
  hoi   c2s_roleserver_getrolelist : pDataBuffer[0] = so nhan vat toi da, [1..] = ten tai khoan
  tra   s2c_..._getrolelist_result : pDataBuffer[0] = so nhan vat, [1..] = mang S3DBI_RoleBaseInfo
  (Goddess\ClientNode.cpp:295-341). Chi lay roles[0] => moi tai khoan dung 1 bot.

Duong sinh bot SOI GUONG CoreServerShell::AddCharacter (:130-153) + PlayerDbLoading
(:155-181) + AddPlayerToWorld (:183-187) - tuc duong dang nhap cua NGUOI THAT o chinh cay
nay, thay vi dich lai KBotManager cua cay khac. It rui ro hon han.

TRAN BOT = 20, gan cung, CO Y.
  KPlayerSet::FindFree lay tu DAU danh sach ro (KPlayerSet.cpp:161), ma danh sach dung tu
  MAX_PLAYER-1 xuong 1 va chen o DAU (:59-62 + KLinkArray.cpp:55-58) => bot chiem khe THAP
  va day nguoi choi that len chi so cao. Ma KSOServer::SavePlayerData chan
  nIndex > m_nMaxPlayer (=450+20=470; KSOServer.cpp:3190, 3173, 3284, 2460) => nguoi that bi
  day qua 470 se KHONG BAO GIO DUOC LUU. Giu tran bot nho la cach vo hieu hoa lo hong do MA
  KHONG dung mot dong nao cua duong nguoi choi that.

Da chan san min ipassrole == 214519 ngay truoc khi nap blob (KPlayer.cpp:6768 goi
g_pServer->Release() roi dong 6776 dung con tro NULL -> giet tang mang may chu).

=== 4 cua giao tiep (Buoc 3-4) ===
CoreServerShell.h: them SSOI_PBOT_SET_SENDER / SPAWN / ROLELIST_RES / ROLEDATA_RES /
REMOVE_ALL va SGDI_PBOT_IS_BOT, dat NGOAI enum theo dung khuon cac ma JX2 co san (:231-236).
KHONG them ham thuan ao: iCoreServerShell duoc CA S3Server dung
(S3Server\SwordOnLineServer.cpp:192); them ham ao la buoc build lai ca no, quen mot noi la
lech khe vtable - hong ngam khong co loi lien ket.
CoreServerShell.cpp: 4 vo mong trong OperationRequest/GetGameData (khuon y het case
SSOI_TONG_JX2_SYNC o :1012-1015), va PB_Breathe() ngay sau SC_Breathe() trong Breathe().

Build: Server Release|x64 PASS va Client Release|Win32 PASS (0 loi ca hai).
Ma hoa TCVN3 khong doi: KProtocol.h 11063, CoreServerShell.h 756, CoreServerShell.cpp 930.

CON LAI cua dot: noi day KSOServer (cai dat PB_DbSender + dinh tuyen 2 goi tra loi),
va xu ly m_nLixian bi engine tu lat 1 -> 2 (KSOServer.cpp:3450-3467).

+++ Sources/Core/Src/KProtocol.h
@@ -1339,7 +1339,7 @@ struct tagNewDelRoleResponse : public tagDBSelPlayer
 // 2003.05.11
 struct tagDBSyncPlayerInfo : public tagProtoHeader
 {
-	size_t	dataLength;
+	unsigned int	dataLength;
 	char	szData[0];
 };
 
@@ -1367,7 +1367,7 @@ struct tagGuidableInfo : public tagProtoHeader
 	GUID guid;
 	WORD nExtPoint;			//¿ÉÓÃµÄ¸½ËÍµã
 	WORD nChangePoint;		//±ä»¯µÄ¸½ËÍµã
-	size_t	datalength;
+	unsigned int	datalength;
 	char	szData[0];
 
 };
@@ -1608,7 +1608,7 @@ struct tagRegisterFamily : public tagProtoHeader
  */
 struct tagGS2GWSaveRole : public tagProtoHeader
 {
-	size_t	datalength;
+	unsigned int	datalength;
 	BYTE	szData[0];
 };
